feat!: Change default mappings for CLR type decimal
BREAKING CHANGE: The default mapping for values of CLR type decimal was FLOAT64 and it is now Numeric.

+++ apis/Google.Cloud.Spanner.Data/Google.Cloud.Spanner.Data.Tests/SpannerConversionOptionsTests.cs
@@ -26,7 +26,7 @@ namespace Google.Cloud.Spanner.Data.Tests
 
             Assert.True(options.UseDBNull);
             Assert.Equal(SpannerDbType.Float32, options.SingleToConfiguredSpannerType);
-            Assert.Equal(SpannerDbType.Float64, options.DecimalToConfiguredSpannerType);
+            Assert.Equal(SpannerDbType.Numeric, options.DecimalToConfiguredSpannerType);
             Assert.Equal(SpannerDbType.Timestamp, options.DateTimeToConfiguredSpannerType);
             Assert.Equal(typeof(DateTime), options.DateToConfiguredClrType);
         }
